[wmii] Bind key to collect windows to 0:1
Add a script that moves all clients on other screens to column 1 of the
primary screen.
Bind the script to $MODKEY-Shift-BackSpace.

Having this keybinding makes it easier to disconnect an external monitor
with multiple clients on it. It will also collect the now-inaccessible
clients if they were not collected before the monitor was disconnected.