[wmii] Add the current client's properties to lbar
This makes it easy to see which client is selected at a glance, even if
that client is off screen.

Since it's in the lbar, precautions have been taken so that it does not
get interpreted as a tag, although it will still be possible to view the
"~selclient" tag.