Ignore everything
When setting up a bare repository in your home folder, similar to the
way that is suggested in an Atlassian blog post[1], it would be better
to ignore everything and only track what is explicitly given. Ignoring
is better than setting 'status.showUntrackedFiles' to "no" because
untracked files cannot be accidentally included.

[1] https://developer.atlassian.com/blog/2016/02/best-way-to-store-dotfiles-git-bare-repo/