[bash] Add dotfiles alias
This alias facilitates the workflow in which dotfiles are tracked using
a bare Git repo in the home folder.

That workflow is described in the Atlassian Developer blog post "The
best way to store your dotfiles: A bare Git repository" by Nicola
Paolucci.

https://developer.atlassian.com/blog/2016/02/best-way-to-store-dotfiles-git-bare-repo/

This repository is managed in a very similar way, except that everything
is ignored by default so nothing will be added accidentally.